requirements.txt
PyYAML==6.0
suntime==1.2.5
TeslaPy==2.7.0
